+++ .gitignore
@@ -8,3 +8,5 @@
 .externalNativeBuild
 .cxx
 local.properties
+firestore-debug.log
+ui-debug.log
