+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/dependencyinjection/FakeRepositoryModule.kt
@@ -34,7 +34,7 @@ abstract class FakeRepositoryModule {
 
     @Singleton
     @Binds
-    abstract fun provideUserRepo(repo: FakeUserRepository): UserRepository
+    abstract fun provideUserRepo(repo: FakeUserRepository): UserRepositoryInterface
 
     @Singleton
     @Binds
