+++ app/src/main/java/com/github/sdp/ratemyepfl/model/user/User.kt
@@ -1,7 +1,6 @@
 package com.github.sdp.ratemyepfl.model.user
 
 import com.github.sdp.ratemyepfl.auth.ConnectedUser
-import com.google.firebase.firestore.DocumentSnapshot
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -11,21 +10,6 @@ data class User(
     val email: String?,
     val picture: String? = "$uid.jpg"
 ) {
-    companion object {
-
-        const val USERNAME_FIELD = "username"
-        const val EMAIL_FIELD = "email"
-        const val PICTURE_FIELD = "picture"
-
-        fun DocumentSnapshot.toUser(): User {
-            return User(
-                uid = id,
-                username = getString(USERNAME_FIELD),
-                email = getString(EMAIL_FIELD),
-                picture = getString(PICTURE_FIELD)
-            )
-        }
-    }
 
     constructor(user: ConnectedUser) : this(
         uid = user.getUserId()!!,
