+++ app/src/main/java/com/github/sdp/ratemyepfl/viewmodel/UserProfileViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.sdp.ratemyepfl.auth.ConnectedUser
 import com.github.sdp.ratemyepfl.database.Storage
-import com.github.sdp.ratemyepfl.database.UserRepository
+import com.github.sdp.ratemyepfl.database.UserRepositoryInterface
 import com.github.sdp.ratemyepfl.model.ImageFile
 import com.github.sdp.ratemyepfl.model.user.User
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,7 +15,7 @@ import javax.inject.Inject
 class UserProfileViewModel @Inject constructor(
     val currentUser: ConnectedUser,
     val imageStorage: Storage<ImageFile>,
-    val userDatabase: UserRepository
+    val userDatabase: UserRepositoryInterface
     ) : ViewModel() {
 
     private val picture : MutableLiveData<ImageFile?> = MutableLiveData(null)
